% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{convert_to_motif}
\alias{convert_to_motif}
\title{Converts a chqracter to a motif}
\usage{
convert_to_motif(s, null_char = ".")
}
\arguments{
\item{s}{a character}

\item{null_char}{the character for unspecified letters}
}
\value{
a motif
}
\description{
Converts a chqracter to a motif
}
