% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{filter_motif_list}
\alias{filter_motif_list}
\title{Filter motifs based on their length}
\usage{
filter_motif_list(motif_list, k_min)
}
\arguments{
\item{motif_list}{list of motifs}

\item{k_min}{minimum motif length}
}
\value{
A list of motifs
}
\description{
Filter motifs based on their length
}
