% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{find_frequent_motifs}
\alias{find_frequent_motifs}
\title{Find frequent motifs in a set of sequences using the Motif_All algorithm}
\usage{
find_frequent_motifs(x, N_support = NULL, support = 0.05, k_max = NULL,
  central_letter = NULL, showProgress = TRUE)
}
\arguments{
\item{x}{a vector of sequences (with a fixed length). Defined as the background set}

\item{N_support}{minimum support of motifs in the foreground set (number of counts). Has priority over \code{support}}

\item{support}{minimum support of motifs in the foreground set (frequence)}

\item{k_max}{maximum size of motifs (set to NULL to ignore)}

\item{central_letter}{search only for motifs that have the given character in central position}

\item{showProgress}{show progress bar in console}
}
\value{
a list of motifs
}
\description{
Find frequent motifs in a set of sequences using the Motif_All algorithm
}
