% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{get_motif_score}
\alias{get_motif_score}
\title{Compute the enrichment score for a list of motifs in the foreground set as compared to the background set}
\usage{
get_motif_score(x, idx_select, motif_list)
}
\arguments{
\item{x}{a vector of sequences (with a fixed length). Defined as the background set}

\item{idx_select}{Indexes of foreground sequences in \code{x}. Note that the foreground set must be a subset of the background set.}

\item{motif_list}{list of motifs}
}
\value{
A data frame summarizing motif enrichment analysis in foreground set
}
\description{
Compute the enrichment score for a list of motifs in the foreground set as compared to the background set
}
