% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{get_unique_motif_score}
\alias{get_unique_motif_score}
\title{Compute the enrichment score for a motif in the foreground set as compared to the background set}
\usage{
get_unique_motif_score(df, idx_select, motif, match_central_letter = TRUE)
}
\arguments{
\item{df}{a data frame corresponding to set of background sequences as obtained by calling \code{sequence_to_df}. 
Could also be a vector of background sequences (with a fixed length) but will be slower.}

\item{idx_select}{Indexes of foreground sequences in \code{x}. Note that the foreground set must be a subset of the background set.}

\item{motif}{a motif}

\item{match_central_letter}{keep only sequences that have the same central letter as the motif}
}
\value{
A data frame summarizing motif enrichment analysis in foreground set
}
\description{
Compute the enrichment score for a motif in the foreground set as compared to the background set
}
