% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{match_motif_df}
\alias{match_motif_df}
\title{Match a motif to a set of sequences represented as a data frame}
\usage{
match_motif_df(df, motif)
}
\arguments{
\item{df}{a data frame corresponding to a set of sequences (with a fixed length).}

\item{motif}{a motif}
}
\value{
vector of matching indexes
}
\description{
Match a motif to a set of sequences represented as a data frame
}
