% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{match_motif_list_to_sequences}
\alias{match_motif_list_to_sequences}
\title{Match a list of motifs to a set of sequences}
\usage{
match_motif_list_to_sequences(x, motif_list, showProgress = TRUE)
}
\arguments{
\item{x}{a vector of sequences (with a fixed length).}

\item{motif_list}{a list of motifs}

\item{showProgress}{show progress bar in console}
}
\value{
list of with matching indexes
}
\description{
Match a list of motifs to a set of sequences
}
