% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{match_motif_to_sequences}
\alias{match_motif_to_sequences}
\title{Match a motif to a set of sequences}
\usage{
match_motif_to_sequences(x, motif)
}
\arguments{
\item{x}{a vector of sequences (with a fixed length).}

\item{motif}{a motif}
}
\value{
vector of matching indexes
}
\description{
Match a motif to a set of sequences
}
