% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{match_sequence_to_motifs}
\alias{match_sequence_to_motifs}
\title{Match a sequence to a list of motifs}
\usage{
match_sequence_to_motifs(x, motif_list)
}
\arguments{
\item{x}{a sequence}

\item{motif_list}{list of motifs}
}
\value{
vector of matching indexes
}
\description{
Match a sequence to a list of motifs
}
