% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{motif}
\alias{motif}
\title{Constructor for objects of class "motif"}
\usage{
motif(positions = NULL, letters = NULL, size = 1)
}
\arguments{
\item{positions}{a vector of integers specifying positions of letters in the motif}

\item{letters}{a vector of single characters (letters)}

\item{size}{the total number of characters (including unspecified letters)}
}
\value{
a motif
}
\description{
Constructor for objects of class "motif"
}
