% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{motif_to_string}
\alias{motif_to_string}
\title{Converts a motif to a character string}
\usage{
motif_to_string(x)
}
\arguments{
\item{x}{a motif}
}
\value{
a character string
}
\description{
Converts a motif to a character string
}
