% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{MotifAll}
\alias{MotifAll}
\title{Analysis of motifs in protein sequences}
\usage{
MotifAll(x, idx_select, support = 0.05, k_min = 3, k_max = NULL,
  central_letter = NULL)
}
\arguments{
\item{x}{a vector of sequences}

\item{support}{the minimum support of motifs}

\item{k_min}{the minimum size of motifs}
}
\value{
a vector of motifs
}
\description{
This package implements the Motif_All algorithm
}
\examples{

#load data :
data(df_motif, package = "MotifAll")
site_residue="S"
df <- df_motif[df_motif$Residue == site_residue, ]

x <- as.character(df$Sequence)
idx_select <- which(!is.na(df$Cluster))

save file for a compaison with the java implementation of MotifAll
write.table(x, "./comp/data_bg.txt", row.names = FALSE, col.names = FALSE, quote = FALSE)
write.table(x[idx_select], "./comp/data_fg.txt", row.names = FALSE, col.names = FALSE, quote = FALSE)

res <- MotifAll(x, idx_select, support = 0.05, k_min = 3, central_letter = site_residue)

}
\author{
Guillaume Voisinne
}
