% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{print_motifs}
\alias{print_motifs}
\title{Print a list motif as a vector of characters}
\usage{
print_motifs(motif_list)
}
\arguments{
\item{motif_list}{a list of motifs}
}
\value{
a vector of characters
}
\description{
Print a list motif as a vector of characters
}
