% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{print_single_motif}
\alias{print_single_motif}
\title{Print a motif as a character string}
\usage{
print_single_motif(motif, n_letters)
}
\arguments{
\item{motif}{a motif}

\item{n_letters}{length of the output character string}
}
\value{
a character string
}
\description{
Print a motif as a character string
}
