% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{score_motif_enrichment}
\alias{score_motif_enrichment}
\title{Compute the enrichment score for a list of motifs in the foreground set as compared to the background set}
\usage{
score_motif_enrichment(x, idx_select, motif, central_letter = NULL)
}
\arguments{
\item{x}{a vector of sequences (with a fixed length). Defined as the background set}

\item{idx_select}{Indexes of foreground sequences in \code{x}. Note that the foreground set must be a subset of the background set.}

\item{motif}{list of motifs}

\item{central_letter}{keep only sequences that have the given character in central position}
}
\value{
A data frame summarizing motif enrichment analysis in foreground set
}
\description{
Compute the enrichment score for a list of motifs in the foreground set as compared to the background set
}
