% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{sequence_to_df}
\alias{sequence_to_df}
\title{Converts a set of sequences to a dataframe}
\usage{
sequence_to_df(x)
}
\arguments{
\item{x}{a vector of sequences (with a fixed length)}
}
\value{
a data frame with letter positions as columns and sequences as rows
}
\description{
Converts a set of sequences to a dataframe
}
