% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{get_unique_motifs}
\alias{get_unique_motifs}
\title{Get unique motifs from a list of motifs}
\usage{
get_unique_motifs(motif_list, n_letters)
}
\arguments{
\item{motif_list}{a list of motifs}

\item{n_letters}{maximum number of letters in a motif}
}
\value{
a list of motifs
}
\description{
Get unique motifs from a list of motifs
}
