% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{match_sequence_to_motifs}
\alias{match_sequence_to_motifs}
\title{Match a sequence to a list of motifs}
\usage{
match_sequence_to_motifs(motif_list, x)
}
\arguments{
\item{motif_list}{list of motifs}

\item{x}{a sequence}
}
\value{
vector of matching indexes
}
\description{
Match a sequence to a list of motifs
}
