% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{MotifAll}
\alias{MotifAll}
\title{Analysis of motifs in protein sequences}
\usage{
MotifAll(x, support = 0.05, k_min = 4)
}
\arguments{
\item{x}{a vector of sequences}

\item{support}{the minimum support of motifs}

\item{k_min}{the minimum size of motifs}
}
\value{
a vector of motifs
}
\description{
This package implements the Motif_All algorithm
}
\examples{

#load data :
x <- as.character(df_motif$Sequence[!is.na(Tsite$Cluster) & Tsite$Residue == "Y"])
motifs <- MotifAll(x, support = 0.05, k_min = 4)

}
\author{
Guillaume Voisinne
}
