% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{MotifAll}
\alias{MotifAll}
\title{Analysis of motifs in protein sequences}
\usage{
MotifAll(x, idx_select, support = 0.05, k_min = 4, k_max = NULL)
}
\arguments{
\item{x}{a vector of sequences}

\item{support}{the minimum support of motifs}

\item{k_min}{the minimum size of motifs}
}
\value{
a vector of motifs
}
\description{
This package implements the Motif_All algorithm
}
\examples{

#load data :
data(df_motif, package = "MotifAll")
idx_select <- which(!is.na(df_motif$Cluster) & df_motif$Residue == "Y")
x <- as.character(df_motif$Sequence)
motifs <- MotifAll(x, idx_select, support = 0.01, k_min = 4)

}
\author{
Guillaume Voisinne
}
