% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifAll.R
\name{MotifAll}
\alias{MotifAll}
\title{Analysis of motifs in protein sequences}
\usage{
MotifAll(x, idx_select, N_support = NULL, support = 0.05, signif = 0.05,
  k_min = 3, k_max = NULL, central_letter = NULL)
}
\arguments{
\item{x}{a vector of sequences (with a fixed length). Defined as the background set}

\item{idx_select}{Indexes of foreground sequences in \code{x}. Note that the 
foreground set must be a subset of the background set.}

\item{N_support}{minimum support of motifs in the foreground set (number of counts). 
Has priority over \code{support}}

\item{support}{minimum support of motifs in the foreground set (frequence)}

\item{signif}{significance level of output motifs (p-value associated with the Z-score)}

\item{k_min}{minimum size of motifs}

\item{k_max}{maximum size of motifs (set to NULL to ignore)}

\item{central_letter}{search only for motifs that have the given character in central position}
}
\value{
A list including the following elements :

\code{score} : data frame with the motifs identified along with scores and counts

\code{idx_match_bckg} : list of vectors containing the matching indexes of a given motif in the set of sequences x

\code{idx_match_sample} : list of vectors containing the matching indexes of a given motif in the set of sequences x also part of the foreground set
}
\description{
This package implements the Motif_All algorithm
}
\examples{
#load data :
dir<- system.file("extdata", package = "MotifAll")
path <- paste(dir, "/phospho_site_annotated_plus_motifs.txt", sep="")
df_motif <- read.csv(path, sep="\\t")
#data(df_motif, package = "MotifAll")
psite <- "S"

idx_residue <- which(substr(as.character(df_motif$Sequence),8,8) == psite)
df <- df_motif[idx_residue, ]

x <- as.character(df$Sequence)
idx_select <- which(!is.na(df$Cluster))

res <- MotifAll(x, idx_select, support = 0.05, signif = 1e-4, k_min = 1, central_letter = psite)

}
\author{
Guillaume Voisinne
}
